+++ omar qt/admin .cpp
@@ -81,68 +81,8 @@ void  QtWidgetsApplication3::on_change_users_clicked()
 
 }
 
-void  QtWidgetsApplication3::on_admin_mainmenu_clicked()
-{
-    ui->stackedWidget->setCurrentWidget(ui->admin);
-}
 
 
-void QtWidgetsApplication3::on_pushButton_10_clicked() {
-
-
-    connect(ui->listWidget_subscriptions, &QListWidget::itemClicked, this, [=](QListWidgetItem* item) {
-        int i = item->data(Qt::UserRole).toInt();
-        chosenSubscriptionIndex = i;
-
-        QString details;
-        details += "📦 Plan Name: " + QString::fromStdString(arr_subscription[i].plan_name) + "\n";
-
-        if (arr_subscription[i].fixed == 'y') {
-            details += "\n🟦 Fixed Subscription:\n";
-
-            if (arr_subscription[i].month_count > 0) {
-                details += "\n📅 Monthly Plans:\n";
-                for (int j = 0; j < arr_subscription[i].month_count; j++) {
-                    details += "➤ " + QString::number(arr_subscription[i].month_sub[j].duration) + " month(s), "
-                        + QString::number(arr_subscription[i].month_sub[j].no_of_trips) + " trips\n";
-                    for (int k = 0; k < arr_subscription[i].month_sub[j].zone_num; k++) {
-                        details += "   • Zone " + QString::number(k + 1) + ": "
-                            + QString::number(arr_subscription[i].month_sub[j].zonesPrice[k]) + " LE\n";
-                    }
-                }
-            }
-
-            if (arr_subscription[i].year_count > 0) {
-                details += "\n📅 Yearly Plans:\n";
-                for (int j = 0; j < arr_subscription[i].year_count; j++) {
-                    details += "➤ " + QString::number(arr_subscription[i].year_sub[j].duration) + " year(s), "
-                        + QString::number(arr_subscription[i].year_sub[j].no_of_trips) + " trips\n";
-                    for (int k = 0; k < arr_subscription[i].year_sub[j].zone_num; k++) {
-                        details += "   • Zone " + QString::number(k + 1) + ": "
-                            + QString::number(arr_subscription[i].year_sub[j].zonesPrice[k]) + " LE\n";
-                    }
-                }
-            }
-        }
-        else {
-            details += "\n🟨 Wallet Subscription:\n";
-            details += "• Add balance in multiples of: " + QString::number(arr_subscription[i].wallet_sub.fund_multiple) + " LE\n";
-            details += "• Maximum card balance: " + QString::number(arr_subscription[i].wallet_sub.card_balance) + " LE\n";
-            details += "• Zones and Prices:\n";
-            for (int k = 0; k < arr_subscription[i].wallet_sub.zone_num; k++) {
-                details += "   • Zone " + QString::number(k + 1) + ": "
-                    + QString::number(arr_subscription[i].wallet_sub.zonesPrice[k]) + " LE\n";
-            }
-          
-        }
-
-        if (!arr_subscription[i].notes.empty()) {
-            details += "\n📝 Notes:\n" + QString::fromStdString(arr_subscription[i].notes);
-        }
-
-        ui->label_subscription_details->setPlainText(details);  // For QTextEdit
-        });
-}
 
 /// for try run
