+++ omar qt/complains.txt
@@ -3,16 +3,3 @@ The sequence of data is (name---> id ----> complaints)
 omar ID1   NO COMPLAINTS
 alol ID2   NO COMPLAINTS
 alol2 ID3 NO COMPLAINTS
-  NO COMPLAINTS
-  NO COMPLAINTS
-  NO COMPLAINTS
-  NO COMPLAINTS
-    NO COMPLAINTS
-    NO COMPLAINTS
-    NO COMPLAINTS
-    NO COMPLAINTS
-  NO COMPLAINTS
-  NO COMPLAINTS
-  NO COMPLAINTS
-  NO COMPLAINTS
-  NO COMPLAINTS
