+++ omar qt/data of users.txt
@@ -1,19 +1,6 @@
 information of person is in this sequence:(name --> id --> role --> balance --> password --> email)
-16
+3
 
 omar ID1 0 50 12345678 omar@gmail.com
-alol ID2 0 0 123456789 aly@gmail.com
-alol2 ID3 0 0 12345678 alol@gmail.com
-  0 0 741852963 3mooor@gmail.com
-  0 0 96385274 haya@gamail.com
-  0 0 88888888888888 ppp@gmail.com
-  0 0 ;;;;;;;;;;;;;;;;;;;;;;;;;;; kkk@gmail.com
-  0 0 0 0
-  0 0 0 0
-  0 0 0 0
-  0 0  
-  0 0  
-  0 0  
-  0 0  
-  0 0  
-  0 0  
+alol ID2 0 500000 123456789 aly@gmail.com
+alol2 ID3 0 10006 12345678 alol@gmail.com
