+++ omar qt/dependence.h
@@ -13,7 +13,6 @@
 #include <utility>
 #include <sstream>
 #include<QString>
-#include<string>
 #include<QMessageBox>
 #include<QDebug>
 #include <QDate>
