+++ omar qt/history rides.txt
@@ -3,3 +3,4 @@ ID1     Ring_Road.     Cairo_University.     10     06/05/2025
 ID2     Wadi_Hof.     Kotsika.     8     24/06/2025
 ID2     Helwan_University.     Saad_Zaghloul.     8     25/06/2025
 ID2     Helwan.     El-Zahraa.     8     27/06/2025
+ID2     Helwan.     El-Ma'sara.     8     27/06/2025
