+++ omar qt/rides .cpp
@@ -242,3 +242,4 @@ void  QtWidgetsApplication3::on_confirmride_clicked()
 
 
 }
+
