+++ omar qt/subsc.txt
@@ -1,3 +1,3 @@
 n omar ID1 0 0 0 0 0
-n alol ID2 Smart_wallet. 0 0 6/5/2025 492
-y alol2 ID3 General_pass. year 1 730 728 2 4000 6/5/2025 6/5/2026
+n alol ID2 Smart_wallet. 0 0 6/5/2025 484
+y alol2 ID3 General_pass. year 1 730 728 2 4000 6/5/2025 27/6/2026
